Create a yabf version of configfile modules deploy
