+++ src/main.js
@@ -3,6 +3,8 @@ export * from './core/term-application'
 export * from './commands/scripts-list.command'
 export * from './commands/scripts-run.command'
 export * from './commands/modules-list.command'
+export * from './commands/modules-deploy.command'
 
 export * from './services/file.service'
-export * from './services/exec'
+export * from './services/exec/exec.service'
+export * from './services/deploy/deploy.service'
